fix(linker): don't symlink on Windows (#1253)
1) FAT (FAT32, exFAT, etc.) file systems doesn't support symlinks.
2) You need administrator access in order to create symlinks.

This was working for me locally because I'm running a NTFS fs with
"Developer Mode" enabled in Windows 10, which bypasses administrator
requirement for mklink.

Instead we uv_fs_rename() links. This should effectively be the same,
but will require special handling down the road if/when for example
multiple package versions can be installed at the same time. When
unlinking a package the renamed files should me moved back into the
package installation directory, instead of simply being unlinked (i.e.
removed).

Closes #1251.