feat(pypi): fallback to virtualenv if venv fails
On Ubuntu 22.04 when the `python3-venv` package is not installed:

* Importing the `venv` module will succeed.
* Running `python3 -m venv --help` will output help.
* Running `python3 -m venv $VENV_DIR` fails with the following output:

```console
$ python3 -m venv $VENV_DIR
The virtual environment was not created successfully because ensurepip is not
available.  On Debian/Ubuntu systems, you need to install the python3-venv
package using the following command.

    apt install python3.10-venv

You may need to use sudo with that command.  After installing the python3-venv
package, recreate your virtual environment.

Failing command: /root/venv/bin/python3
```

When running in a rootless environment installing the `python3-venv`
package is not tenable. However, the `virtualenv` package can be
installed via `pip` in the users home directory.

This patch checks the result of `python3 -m venv $VENV_DIR`, if that
fails then it will fallback to calling `python3 -m virtualenv $VENV_DIR`
instead.